scraper/requirements.txt
# Python scraper dependencies
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
playwright>=1.40.0
